Separate React contexts from providers
The eslint-plugin-react-refresh bump to 0.4.26 (an incidental
side effect of the npm vulnerability lockfile update) made the
only-export-components rule flag files that export a context
next to a provider component, breaking the zero-warning lint
gate on eight core contexts. Each context now lives in its own
module, restoring Fast Refresh and a clean lint.
