+++ src/Geopilot.Frontend/src/auth/index.ts
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import { ApiAuthConfigurationContext } from "./apiAuthConfigurationContext";
-import { GeopilotAuthContext } from "./geopilotAuthComponent";
+import { GeopilotAuthContext } from "./geopilotAuthContext";
 import { UserContext } from "./userContext";
 
 export const useGeopilotAuth = () => useContext(GeopilotAuthContext);
