+++ src/Geopilot.Frontend/src/auth/userProvider.tsx
@@ -1,0 +1,40 @@
+import { FC, PropsWithChildren, useCallback, useEffect, useState } from "react";
+import { AuthContextProps, useAuth } from "react-oidc-context";
+import { ApiError, User } from "../api/apiInterfaces";
+import useFetch from "../hooks/useFetch.ts";
+import { UserContext } from "./userContext";
+
+export const UserProvider: FC<PropsWithChildren> = ({ children }) => {
+  const [user, setUser] = useState<User | null>();
+  const auth = useAuth();
+  const { fetchApi } = useFetch();
+
+  const fetchUserInfo = useCallback(
+    (auth: AuthContextProps) => {
+      fetchApi<User>("/api/v1/user/self", {
+        headers: {
+          Authorization: `Bearer ${auth.user?.access_token}`,
+        },
+        errorMessageLabel: "userNoPermission",
+      })
+        .then(setUser)
+        .catch(error => {
+          if (error instanceof ApiError && (error.status === 401 || error.status === 403)) {
+            auth.signoutSilent();
+            setUser(null);
+          }
+        });
+    },
+    [fetchApi],
+  );
+
+  useEffect(() => {
+    if (auth?.isAuthenticated) {
+      fetchUserInfo(auth);
+    } else if (auth && !auth.isLoading) {
+      setUser(null);
+    }
+  }, [auth, auth?.isAuthenticated, auth?.isLoading, fetchUserInfo]);
+
+  return <UserContext.Provider value={user}>{children}</UserContext.Provider>;
+};
