+++ src/Geopilot.Frontend/src/components/controlledNavigate/index.ts
@@ -1,4 +1,4 @@
 import { useContext } from "react";
-import { ControlledNavigateContext } from "./controlledNavigateProvider.tsx";
+import { ControlledNavigateContext } from "./controlledNavigateContext";
 
 export const useControlledNavigate = () => useContext(ControlledNavigateContext);
